requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-chroma==0.1.2
langchain-text-splitters==0.2.0
chromadb==0.4.24
openai==1.59.7
pymupdf==1.24.2
pillow==10.3.0
python-multipart==0.0.9
markdown==3.6
beautifulsoup4==4.12.3
python-docx==1.1.2
python-pptx==0.6.23
pandas==2.2.2
